feat: gcc 13.3.0, newlib 4.5.0, binutils 2.44 - A new patch has been added to disable cunroll_grow_size, introduced in https://github.com/gcc-mirror/gcc/commit/60bd3f20baebeeddd60f8a2b85927e7da7c6016e for GCC 10.2.0 which breaks compiled bins (idk why) - Updated usb_simple_request to match typings from usb_std_request - Aligned usbctrl_set_rol with these typing patterns - Crude patch for libfat to allow for the newlib changes

+++ libxenon/drivers/usb/usbd.h
@@ -227,7 +227,7 @@ int usb_get_config_descriptor(usbdev_t *dev,usb_config_descr_t *dscr,int idx,int
 int usb_get_device_status(usbdev_t *dev,usb_device_status_t *status);
 int usb_set_configuration(usbdev_t *dev,int config);
 int usb_open_pipe(usbdev_t *dev,usb_endpoint_descr_t *epdesc);
-int usb_simple_request(usbdev_t *dev,uint8_t reqtype,int bRequest,int wValue,int wIndex);
+int usb_simple_request(usbdev_t *dev,uint8_t reqtype,uint8_t bRequest,uint16_t wValue,uint16_t wIndex);
 void usb_complete_request(usbreq_t *ur,int status);
 int usb_get_device_descriptor(usbdev_t *dev,usb_device_descr_t *dscr,int smallflg);
 int usb_set_ep0mps(usbdev_t *dev,int mps);
